Run a subset of public repos during normal tests
