--- .github/workflows/bento.yml
@@ -1,14 +1,0 @@
-jobs:
-  bento:
-    name: Check
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v1
-    - id: bento
-      name: Bento
-      uses: returntocorp/bento-action@v1
-      with:
-        acceptTermsWithEmail: russell.r.cohen@gmail.com
-name: Bento
-'on':
-- pull_request
