+++ semgrep/tests/e2e/test_semgrep_rules_repo.py
@@ -3,9 +3,6 @@ import subprocess
 import pytest
 
 
-pytestmark = pytest.mark.qa
-
-
 def test_semgrep_rules_repo(run_semgrep_in_tmp):
     subprocess.check_output(
         ["git", "clone", "--depth=1", "https://github.com/returntocorp/semgrep-rules"]
